modify build scripts and add extra targets

+++ build.sh
@@ -1,32 +1,33 @@
 #!/bin/bash
-DL="https://github.com/openipc/firmware/releases/download/latest"
+DL="https://github.com/OpenIPC/firmware/releases/download/toolchain/toolchain"
 
-if [[ "$1" == *"star6b0" ]]; then
-	CC=cortex_a7_thumb2_hf-gcc13-musl-4_9
-	GCC=$PWD/toolchain/$CC/bin/arm-linux-gcc
-elif [[ "$1" == *"star6e" ]]; then
-	CC=cortex_a7_thumb2_hf-gcc13-glibc-4_9
-	GCC=$PWD/toolchain/$CC/bin/arm-linux-gcc
-elif [[ "$1" == *"goke" ]]; then
-	CC=cortex_a7_thumb2-gcc13-musl-4_9
-	GCC=$PWD/toolchain/$CC/bin/arm-linux-gcc
-elif [[ "$1" == *"hisi" ]]; then
-	CC=cortex_a7_thumb2-gcc13-musl-4_9
-	GCC=$PWD/toolchain/$CC/bin/arm-linux-gcc
-else
-	echo "Usage: $0 [goke|hisi|star6b0|star6e|]"
+if [ "$#" -ne 1 ]; then
+	echo "Usage: $0 [goke|hisi|hi3536|star6b0|star6e|star6c|native]"
 	exit 1
 fi
 
-OUT=$PWD/release/$1
+if [[ "$1" == *"star6b0" ]]; then
+	CC=sigmastar-infinity6b0
+elif [[ "$1" == *"star6e" ]]; then
+	CC=sigmastar-infinity6e
+elif [[ "$1" == *"star6c" ]]; then
+	CC=sigmastar-infinity6c
+elif [[ "$1" == *"goke" ]]; then
+	CC=goke-gk7205v200
+elif [[ "$1" == *"hisi" ]]; then
+	CC=hisilicon-hi3516ev200
+elif [[ "$1" == *"hi3536" ]]; then
+	CC=hisilicon-hi3536dv100
+fi
 
-mkdir -p $OUT
-
-if [ ! -e toolchain/$CC ]; then
-	wget -c -q --show-progress $DL/$CC.tgz -P $PWD
-	mkdir -p toolchain/$CC
-	tar -xf $CC.tgz -C toolchain/$CC --strip-components=1 || exit 1
-	rm -f $CC.tgz
+if [[ "$1" != *"native"* ]]; then
+	if [ ! -e toolchain/$CC ]; then
+		wget -c -q --show-progress $DL.$CC.tgz -P $PWD
+		mkdir -p toolchain/$CC
+		tar -xf toolchain.$CC.tgz -C toolchain/$CC --strip-components=1 || exit 1
+		rm -f $CC.tgz
+	fi
+	OUT=msposd_$1
 fi
 
 if [ ! -e firmware ]; then
@@ -35,14 +36,23 @@ fi
 
 if [ "$1" = "goke" ]; then
 	DRV=$PWD/firmware/general/package/goke-osdrv-gk7205v200/files/lib
-	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC OUTPUT=$OUT $1
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
 elif [ "$1" = "hisi" ]; then
 	DRV=$PWD/firmware/general/package/hisilicon-osdrv-hi3516ev200/files/lib
-	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC OUTPUT=$OUT $1
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
+elif [ "$1" = "hi3536" ]; then
+	DRV=$PWD/firmware/general/package/hisilicon-osdrv-hi3536dv100/files/lib
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
 elif [ "$1" = "star6b0" ]; then
 	DRV=$PWD/firmware/general/package/sigmastar-osdrv-infinity6b0/files/lib
-	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC OUTPUT=$OUT $1
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
 elif [ "$1" = "star6e" ]; then
 	DRV=$PWD/firmware/general/package/sigmastar-osdrv-infinity6e/files/lib
-	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC OUTPUT=$OUT $1
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
+elif [ "$1" = "star6c" ]; then
+	DRV=$PWD/firmware/general/package/sigmastar-osdrv-infinity6c/files/lib
+	make -B CC=$GCC DRV=$DRV TOOLCHAIN=$PWD/toolchain/$CC $1
+else
+	DRV=$PWD
+	make DRV=$DRV $1
 fi
